const correctness

+++ worker/src/RTC/RtpStream.cpp
@@ -274,7 +274,7 @@ namespace RTC
 		// clang-tidy "thinks" that this can lead to division by zero but we are
 		// smarter.
 		// NOLINTNEXTLINE(clang-analyzer-core.DivideZero)
-		this->score = static_cast<uint8_t>(std::round(static_cast<float>(totalScore) / samples));
+		this->score = static_cast<uint8_t>(std::round(static_cast<double>(totalScore) / samples));
 
 		// Call the listener if the global score has changed.
 		if (this->score != previousScore)
